+++ cpp_conan/conanfile.py
@@ -9,7 +9,7 @@ class ExampleConan(ConanFile):
     version = "0.1.0-snapshot"
     settings = "os", "arch", "compiler", "build_type"
 
-    #requires = "boost/1.83.0"
+    requires = "boost/1.83.0"
     tool_requires = "doxygen/1.9.4"
     #test_requires = "boost/1.83.0"