+++ cpp_conan/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog
+# Changelog    {#changelog_md}
 
 All notable changes to this project will be documented in this file.
 
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * doxygen documentation
+* example executable